Add public benchmark harness plugin
