+++ tests/plugins/benchmark_harness/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the benchmark harness plugin."""
